Some very basic theming support - fix #1
This is still very limited.
It introduces a basic plugin system, so it's easy to add new mergers
supporting new software (see BaseMerger.py for a short description on
what to do and mergers/i3.py for an example).

Some basic backup is implemented, but it's very rudimentary (see
themes.py)

The structure of a theme is very simple.
The name of it is the name of its directory,
which includes one directory for each software it supports
containing the config files to use.

Example structure:
    blue/
    --->i3/
    ------->config

This would create a theme »blue« including only a config file for i3.