Rewrite merger system a bit
Mergers can now declare support for multiple softwares making more
general mergers possible. They can also state a priority, so one merger
can basically claim support for every software but with a negative
priority so it doesn't get in the way of more specific mergers.

Also introduces a better backup system, a file is now copied to the
originals theme the first time we try to patch it. Afterwards changes
have to be made in the originals theme instead of the actual config
files.

TODO: regex support for software/filenames (e.g. for a general
"replace-file-merger")